Refactor Solana coin handling and improve custom token integration
- Removed the old findSolanaCoins implementation and replaced it with a new structure that utilizes getSolanaToken for better modularity.
- Introduced getSolanaCustomToken to streamline the process of fetching Solana token details.
- Updated the custom token resolver to integrate the new Solana functionality, enhancing overall token management.
