+++ core/ui/chain/coin/addCustomToken/core/getCustomToken.ts
@@ -3,15 +3,14 @@ import { getChainKind } from '@core/chain/ChainKind'
 import { CustomTokenEnabledChainKind } from './chains'
 import { CustomTokenResolver } from './CustomTokenResolver'
 import { getEvmCustomToken } from './evm'
+import { getSolanaCustomToken } from './solana'
 
 const handlers: Record<
   CustomTokenEnabledChainKind,
   CustomTokenResolver<any>
 > = {
   evm: getEvmCustomToken,
-  solana: async () => {
-    throw new Error('Not implemented')
-  },
+  solana: getSolanaCustomToken,
 }
 
 export const getCustomToken: CustomTokenResolver = async input => {
